Move CollectionConverters to javaapi
