+++ src/compiler/scala/tools/nsc/PipelineMain.scala
@@ -25,7 +25,7 @@ import javax.tools.ToolProvider
 import scala.collection.{immutable, mutable}
 import scala.concurrent._
 import scala.concurrent.duration.Duration
-import scala.jdk.CollectionConverters.Ops._
+import scala.jdk.CollectionConverters._
 import scala.reflect.internal.pickling.PickleBuffer
 import scala.reflect.internal.util.FakePos
 import scala.reflect.io.RootPath
