+++ src/library/scala/collection/convert/JavaCollectionWrappers.scala
@@ -17,7 +17,7 @@ package convert
 import java.util.{concurrent => juc}
 import java.{lang => jl, util => ju}
 
-import scala.jdk.CollectionConverters.Ops._
+import scala.jdk.CollectionConverters._
 import scala.language.higherKinds
 
 /** Wrappers for exposing Scala collections as Java collections and vice-versa */
