+++ test/files/jvm/strictfp/Test_2.scala
@@ -2,7 +2,7 @@ import scala.tools.asm.Opcodes
 import scala.tools.asm.tree._
 import scala.tools.partest.BytecodeTest
 
-import scala.jdk.CollectionConverters.Ops._
+import scala.jdk.CollectionConverters._
 
 object Test extends BytecodeTest {
