+++ test/files/jvm/t8582.scala
@@ -1,5 +1,5 @@
 import scala.tools.partest.BytecodeTest
-import scala.jdk.CollectionConverters.Ops._
+import scala.jdk.CollectionConverters._
 
 package p1 {
   package p2 {
