+++ test/files/run/t7974/Test.scala
@@ -4,7 +4,7 @@ import scala.tools.partest.BytecodeTest
 import scala.tools.nsc.backend.jvm.AsmUtils
 import scala.tools.asm.util._
 import scala.tools.nsc.util.stringFromWriter
-import scala.jdk.CollectionConverters.Ops._
+import scala.jdk.CollectionConverters._
 
 object Test extends BytecodeTest {
   def show: Unit = {
